Ticket #235: Fix deletion of options that contained only media
git-svn-id: svn://suivarro/touchstone/trunk@396 8d9c46d6-3bce-41d5-b1ab-56de0fb72b2f

+++ touchstone/question/edit/dichotomous.php
@@ -144,8 +144,18 @@ if (isset($_POST['submit']) and $_POST['submit'] == 'Correct') {
 
     $stem_changes = false;
     for ($option_no=1; $option_no<=15; $option_no++) {
+      $old_media_deleted = false;
+  
+      if (isset($_POST["delete_media$option_no"]) and $_POST["delete_media$option_no"] == '1') {
+        deleteMedia($_POST["old_option_media$option_no"]);
+        $tmp_option_media = '';
+        $tmp_width = 0;
+        $tmp_height = 0;
+        $old_media_deleted = true;
+      }
+    
       $tmp_option_media = unique_filename($_FILES["option_media$option_no"]['name']);
-      if ($_POST["option_text$option_no"] == '' and $tmp_option_media == '' and $_POST["old_option_text$option_no"] != '') {
+      if (isset($_POST["optionid$option_no"]) and $_POST["optionid$option_no"] != '' and $_POST["option_text$option_no"] == '' and $tmp_option_media == '' and ($_POST["old_option_media$option_no"] == '' or $old_media_deleted)) {
         // Delete operation.
         $changes = true;
         $temp_id = $_POST["optionid$option_no"];
@@ -179,7 +189,7 @@ if (isset($_POST['submit']) and $_POST['submit'] == 'Correct') {
       } elseif ($_POST["option_text$option_no"] != '' or ( isset($_POST["old_option_media$option_no"]) and $_FILES["option_media$option_no"]['name'] != $_POST["old_option_media$option_no"] ) ) {
         // Edit operation.
         if ($_FILES["option_media$option_no"]['name'] != '' and $_FILES["option_media$option_no"]['name'] != $_POST["old_option_media$option_no"]) {
-          if (isset($_POST["old_option_media$option_no"]) and $_POST["old_option_media$option_no"] != '') {
+          if (!$old_media_deleted and isset($_POST["old_option_media$option_no"]) and $_POST["old_option_media$option_no"] != '') {
             deleteMedia ("../media/" . $_POST["old_option_media$option_no"]);
           }
           $tmp_option_media = uploadFile("option_media$option_no",$tmp_width,$tmp_height);
@@ -192,11 +202,7 @@ if (isset($_POST['submit']) and $_POST['submit'] == 'Correct') {
           $tmp_option_media = $_POST["old_option_media$option_no"];
           $tmp_width = $_POST["old_option_media_width$option_no"];
           $tmp_height = $_POST["old_option_media_height$option_no"];
-          if (isset($_POST["delete_media$option_no"]) and $_POST["delete_media$option_no"] == '1') {
-            deleteMedia($_POST["old_option_media$option_no"]);
-            $tmp_option_media = '';
-            $tmp_width = 0;
-            $tmp_height = 0;
+          if($old_media_deleted) {
             $stem_changes = true;
             $result = $mysqli->prepare("INSERT INTO track_changes VALUES (NULL, 'Media $option_no',?,$userID,?,'Deleted',NOW(),'Media $option_no')");
             $result->bind_param('is', $_GET['q_id'], $_POST["old_option_media$option_no"]);
@@ -208,7 +214,7 @@ if (isset($_POST['submit']) and $_POST['submit'] == 'Correct') {
         $part_names = array('option_text','option_right_fback','option_wrong_fback','correct');
         foreach($part_names as $section_name) {
           $new_section_name = $section_name . "$option_no";
-          $new_section_name = $_POST[$new_section_name];
+          $new_section_name = (isset($_POST[$new_section_name])) ? $_POST[$new_section_name] : '';
           $old_section_name = 'old_' . $section_name . "$option_no";
           $old_section_name = $_POST[$old_section_name];
           if (trim($new_section_name) != trim($old_section_name)) {
